LayoutTests/imported/w3c/web-platform-tests/web-animations/responsive/to-style-change-expected.txt


FAIL Lengths responsive to style changes assert_equals: expected "calc(40% + 150px)" but got "calc(40% + 100px)"
PASS Percentages responsive to width style changes
PASS Numbers responsive to style changes
